+++ app/src/main/java/com/sandbox/ai/data/model/ImageAnalysisRequest.kt
@@ -1,4 +1,4 @@
-package com.sandbox.ai
+package com.sandbox.ai.data.model
 
 import kotlinx.serialization.Serializable
